Fixes curve generation in HDR mode - Read me
This version uses a secondary final LUT pass to correctly colour grade the results, when the original was set to HDR output.

In this case, I used the version of the LUT effect with the SC Post Effects asset available in the asset store (you'd need to buy your own copy to use this).

This version of the commit allows you to use both curves in the original profile, and tone mapping together - this means this version is about as close as it should be possible to get to the old system within the new one. Internal test content is quite positive.